Create 27. Remove Element
This solution removes all occurrences of a given value 'val' from the array nums in-place and returns the new length of the array. It uses two pointers: i to iterate through the array and v to track the position where the next valid (non-val) element should go. If an element is not equal to val, it's copied to index v, and v is incremented. In the end, v represents the number of elements that are not val, which is returned as the new length.