Create 35. Search Insert Position
This solution finds the index where a given target should be inserted in a sorted array nums to maintain the order. It loops through the array and returns the index of the first number that is greater than or equal to the target. If no such number is found (i.e., the target is larger than all elements), it returns nums.length, meaning the target should be inserted at the end.